Add update functionality to the CLI.
